refactor SetTagProperties
